Gosec 2.22 fixes

+++ models/snapshotrule.go
@@ -27,7 +27,7 @@ type SnapshotRule struct {
 	TimeOfDay        types.String `tfsdk:"time_of_day"`
 	TimeZone         types.String `tfsdk:"timezone"`
 	DaysOfWeek       types.List   `tfsdk:"days_of_week"`
-	DesiredRetention types.Int64  `tfsdk:"desired_retention"`
+	DesiredRetention types.Int32  `tfsdk:"desired_retention"`
 	IsReplica        types.Bool   `tfsdk:"is_replica"`
 	NASAccessType    types.String `tfsdk:"nas_access_type"`
 	IsReadOnly       types.Bool   `tfsdk:"is_read_only"`
